tarkov realese waitnig

+++ internal/service/service.go
@@ -3,11 +3,14 @@ package service
 import (
 	"cards/internal/model"
 	"context"
+	"fmt"
 	"log/slog"
 )
 
+// интерфейс для связи с репозиторий
 type Repo interface {
-	AddCard(ctx context.Context, card model.MindCard) error
+	AddCard(ctx context.Context, card *model.MindCard) error
+	UptadeCardDescription(ctx context.Context, updt []string)
 	DeleteCard(ctx context.Context, title string) error
 }
 
@@ -23,15 +26,37 @@ func New(repo Repo) *Service {
 
 func (s *Service) AddCard(ctx context.Context, title, description, tag string) error {
 	card := model.NewCard(title, description, tag)
-	if err := s.Repo.AddCard(ctx, *card); err != nil {
+
+	if err := s.Repo.AddCard(ctx, card); err != nil {
 		slog.Error("failed to add card",
 			"error", err,
 			"package", "service")
 		return err
 	}
+
 	return nil
 }
 
 func (s *Service) DeleteCard(ctx context.Context, title string) error {
+	if title == "" {
+		return fmt.Errorf("card not exist")
+	}
+
 	return s.Repo.DeleteCard(ctx, title)
+}
+
+func (s *Service) UptadeCardDescription(ctx context.Context, title, description string) error {
+	if title == "" {
+		return fmt.Errorf("nil title")
+	}
+	if description == "" {
+		return fmt.Errorf("nil desc")
+	}
+	//Упросить
+	cardToRepo := []string{title, description}
+
+	s.Repo.UptadeCardDescription(ctx, cardToRepo)
+
+	return nil
+
 }
